refactor: 의미없는 throws 제거
