--- src/main/java/chess/dao/DatabaseInfo.java
@@ -1,11 +1,0 @@
-package chess.dao;
-
-public class DatabaseInfo {
-	final static String SERVER = "localhost:13306";
-	final static String DATABASE = "woowachess";
-	final static String OPTION = "?useSSL=false&serverTimezone=UTC&characterEncoding=utf8";
-	final static String USER_NAME = "root";
-	final static String PASSWORD = "root";
-	final static String PROTOCOL = "jdbc:mysql://";
-	final static String DRIVER = "com.mysql.cj.jdbc.Driver";
-}
